make response generic
